DRF_projects/musicpjt/requirements.txt
asgiref==3.5.2
Django==3.2
django-seed==0.3.1
djangorestframework==3.14.0
Faker==15.1.1
path.py==10.6
python-dateutil==2.8.2
pytz==2022.4
seed==0.11.3
six==1.16.0
sqlparse==0.4.3
toposort==1.7
